Refactor: emotion 객체의 이름을 리팩터링 - Layout을 Wrapper로, SlideControlButton을 StyledSlideControlButton으로 리팩터링
