+++ src/ProductDetail/ProductWrapper.jsx
@@ -6,8 +6,8 @@ import styled from '@emotion/styled';
 
 import PRODUCT_TAG from '../styles/productTag';
 
-import Slide from './Slide';
-import SlideAlbum from './SlideAlbum';
+import Slide from './ProductWrapper/Slide';
+import SlideAlbum from './ProductWrapper/SlideAlbum';
 import ItemInfo from './ItemInfo';
 
 import { updateSlide, setAlbumPosition } from '../utils';
